feat(map): kill player and bomb when obstacle is placed on them

+++ core/src/com/ict1009/pokemanz/entity/Player.java
@@ -283,6 +283,9 @@ public class Player extends Sprite implements ControllerListener, Destoryable {
      * @param delta: 1/fps
      */
     public void update(float delta) {
+        if (destroyed)
+            return;
+
         if (toDestroy && !destroyed) {
             world.destroyBody(body);
             destroyed = true;
